move examples to own folder and fix inventory department description
