+++ src/app/core/core.service.ts
@@ -2,12 +2,12 @@ import { Injectable } from '@angular/core';
 import { IndexedDbService } from '../indexedDb/indexed-db.service';
 import { Directory } from '../shared/models/directory';
 import { Settings } from '../shared/models/settings';
-import { MockPhast, MockPhastSettings } from './mockPhast';
-import { MockPsat, MockPsatCalculator, MockPsatSettings } from './mockPsat';
-import { MockFsat, MockFsatSettings, MockFsatCalculator } from './mockFsat';
-import { MockSsmt, MockSsmtSettings } from './mockSsmt';
-import { MockTreasureHunt, MockTreasureHuntSettings } from './mockTreasureHunt';
-import { MockMotorInventory } from './mockMotorInventoryData';
+import { MockPhast, MockPhastSettings } from '../examples/mockPhast';
+import { MockPsat, MockPsatCalculator, MockPsatSettings } from '../examples/mockPsat';
+import { MockFsat, MockFsatSettings, MockFsatCalculator } from '../examples/mockFsat';
+import { MockSsmt, MockSsmtSettings } from '../examples/mockSsmt';
+import { MockTreasureHunt, MockTreasureHuntSettings } from '../examples/mockTreasureHunt';
+import { MockMotorInventory } from '../examples/mockMotorInventoryData';
 import { BehaviorSubject } from 'rxjs';
 @Injectable()
 export class CoreService {
