+++ src/app/examples/mockTreasureHunt.ts
